♻️ refactor: Update import paths for login and register forms
