+++ frontend/src/components/layouts/navbar/nav-bar.tsx
