+++ frontend/src/hooks/useRegister.ts
@@ -1,7 +1,7 @@
 'use client';
 
 import { useMutation } from '@tanstack/react-query';
-import { api } from '@/utils/api';
+import { api } from '@/lib/api';
 import { RegisterFormData } from '@/schemas/registerSchema';
 
 interface RegisterResponse {
