+++ frontend/src/services/reviewService.ts
@@ -1,5 +1,5 @@
 import { ReviewResponse, Review } from '@/types';
-import { api } from '@/utils/api';
+import { api } from '@/lib/api';
 
 export const reviewService = {
   createReview: async (reviewData: Review) => {
