+++ frontend/src/types/review.ts
@@ -1,0 +1,6 @@
+export interface Review {
+  recipe: string; // Recipe ID
+  reviewer: string; // User ID
+  rating: number;
+  review?: string;
+}
